Move test-specific classes into test-folder
