+++ tests/Parameters/EndMeetingParametersTest.php
@@ -24,8 +24,6 @@ use BigBlueButton\TestCase;
 
 /**
  * @internal
- *
- *
  */
 class EndMeetingParametersTest extends TestCase
 {
