+++ tests/Parameters/HooksDestroyParametersTest.php
@@ -24,8 +24,6 @@ use BigBlueButton\TestCase;
 
 /**
  * @internal
- *
- *
  */
 class HooksDestroyParametersTest extends TestCase
 {
