+++ tests/Parameters/PublishRecordingsParametersTest.php
@@ -24,8 +24,6 @@ use BigBlueButton\TestCase;
 
 /**
  * @internal
- *
- *
  */
 class PublishRecordingsParametersTest extends TestCase
 {
