+++ tests/Responses/GetMeetingInfoResponseTest.php
@@ -22,7 +22,7 @@ namespace BigBlueButton\Responses;
 
 use BigBlueButton\Core\Meeting;
 use BigBlueButton\TestCase;
-use BigBlueButton\Util\Fixtures;
+use BigBlueButton\TestServices\Fixtures;
 
 /**
  * @internal
